Added stderr output as additional callback parameter (#19)
* Added stderr output as additional callback parameter

gpg sends some additional information to stderr. For example the key
verification information during decryption. i
To provide access to this information without redirecting it to stdout
this change adds the "error" as a third parameter to the callback function.
This allows users to get the decrypted content and the additional
information separately if needed.

* added missing test for errors